+++ src/builtins/cd.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 15:16:05 by wnocchi           #+#    #+#             */
-/*   Updated: 2024/08/08 17:52:52 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/09 11:16:50 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,16 +22,16 @@ char	*get_home(t_env *env)
 	res = NULL;
 	while (current)
 	{
-			if (current->key && ft_strcmp(current->key, "HOME") == 0)
-			{
-				if(current->value)
-					res = ft_strdup(current->value);
-				else
-				 	break ;
-				if (!res)
-					return (NULL);
-				return (res);
-			}
+		if (current->key && ft_strcmp(current->key, "HOME") == 0)
+		{
+			if (current->value)
+				res = ft_strdup(current->value);
+			else
+				break ;
+			if (!res)
+				return (NULL);
+			return (res);
+		}
 		current = current->next;
 	}
 	return (ft_err("msh: cd: $HOME not set"), NULL);
@@ -79,7 +79,7 @@ int	ft_cd(char **arg, t_env **env)
 	path = NULL;
 	if (*arg && ft_strcmp(*arg++, "cd") == 0)
 	{
-		if(*arg && chdir(*arg) != -1)
+		if (*arg && chdir(*arg) != -1)
 			return (set_excode(env, 0), 0);
 		pwd = getcwd(NULL, 0);
 		if (!pwd)
