+++ src/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 11:46:46 by wnocchi           #+#    #+#             */
-/*   Updated: 2024/08/08 13:36:09 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/09 11:18:31 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ char	*get_key_env(char *value)
 	int	i;
 
 	i = 0;
-	if(!value)
+	if (!value)
 		return (NULL);
 	if (!is_equal(value) || is_equal(value) == 3)
 		return (ft_strdup(value));
@@ -69,28 +69,6 @@ char	*get_value_env(char *value)
 		return (NULL);
 	else
 		return (ft_substr(value, i + 1, (ft_strlen(value) - i)));
-}
-
-void	add_underscore(t_env *env)
-{
-	t_env	*current;
-	t_env	*tmp;
-	bool	found;
-
-	found = false;
-	current = env;
-	while (current)
-	{
-		if(ft_strcmp(current->full_var, "_=/usr/bin/env") == 0)
-			found = true;
-		current = current->next;
-	}
-	if (found == false)
-	{
-		tmp = no_env("_=/usr/bin/env");
-		add_env_node(&env, tmp);
-		return ;
-	}
 }
 
 int	split_env(t_env **env)
