+++ src/builtins/export.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 10:31:19 by wnocchi           #+#    #+#             */
-/*   Updated: 2024/08/08 13:31:49 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/09 11:24:12 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,232 +14,253 @@
 
 //------------------env_manage.c--------------------
 
-int	free_env(t_env **env)
-{
-	t_env	*current;
-	t_env	*prev;
+// int	free_env(t_env **env)
+// {
+// 	t_env	*current;
+// 	t_env	*prev;
 
-	current = *env;
-	while (current)
-	{
-		prev = current;
-		free(current->value);
-		free(current->full_var);
-		free(current->key);
-		free_tab(current->full_env);
-		current = current->next;
-		free(prev);
-	}
-	free(env);
-	return (0);
-}
+// 	current = *env;
+// 	while (current)
+// 	{
+// 		prev = current;
+// 		free(current->value);
+// 		free(current->full_var);
+// 		free(current->key);
+// 		free_tab(current->full_env);
+// 		current = current->next;
+// 		free(prev);
+// 	}
+// 	free(env);
+// 	return (0);
+// }
 
-int	env_len(t_env *env)
-{
-	int		i;
-	t_env	*current;
+// int	env_len(t_env *env)
+// {
+// 	int		i;
+// 	t_env	*current;
 
-	current = env;
-	i = 0;
-	while (current)
-	{
-		current = current->next;
-		i++;
-	}
-	return (i);
-}
+// 	current = env;
+// 	i = 0;
+// 	while (current)
+// 	{
+// 		current = current->next;
+// 		i++;
+// 	}
+// 	return (i);
+// }
 
-t_env	*ft_envlastnode(t_env *lst)
-{
-	t_env	*current;
+// t_env	*ft_envlastnode(t_env *lst)
+// {
+// 	t_env	*current;
 
-	if (!lst)
-		return (0);
-	current = lst;
-	while (current->next)
-		current = current->next;
-	return (current);
-}
+// 	if (!lst)
+// 		return (0);
+// 	current = lst;
+// 	while (current->next)
+// 		current = current->next;
+// 	return (current);
+// }
 
-void	add_env_node(t_env **lst, t_env *add)
-{
-	if (!lst || !add)
-		return ;
-	if (!*lst)
-		*lst = add;
-	else
-		ft_envlastnode(*lst)->next = add;
-}
+// void	add_env_node(t_env **lst, t_env *add)
+// {
+// 	if (!lst || !add)
+// 		return ;
+// 	if (!*lst)
+// 		*lst = add;
+// 	else
+// 		ft_envlastnode(*lst)->next = add;
+// }
 
-t_env	*create_env_node(char **envp, int i)
-{
-	t_env	*env;
+// t_env	*create_env_node(char **envp, int i)
+// {
+// 	t_env	*env;
 
-	env = ft_calloc(sizeof(t_env), 1);
-	if (!env)
-		return (NULL);
-	env->full_var = ft_strdup(envp[i]);
-	if (!env->full_var)
-		return (NULL);
-	return (env);
-}
+// 	env = ft_calloc(sizeof(t_env), 1);
+// 	if (!env)
+// 		return (NULL);
+// 	env->full_var = ft_strdup(envp[i]);
+// 	if (!env->full_var)
+// 		return (NULL);
+// 	return (env);
+// }
 
 //------------------env_create.c------------------
-t_env	*no_env(char *full_var)
-{
-	t_env	*tmp;
+// t_env	*no_env(char *full_var)
+// {
+// 	t_env	*tmp;
 
-	tmp = ft_calloc(sizeof(t_env), 1);
-		if(!tmp)
-			return (NULL);
-	tmp->full_var = ft_strdup(full_var);
-	return (tmp);
-}
+// 	tmp = ft_calloc(sizeof(t_env), 1);
+// 		if(!tmp)
+// 			return (NULL);
+// 	tmp->full_var = ft_strdup(full_var);
+// 	return (tmp);
+// }
 
+// void	add_underscore(t_env *env)
+// {
+// 	t_env	*current;
+// 	t_env	*tmp;
+// 	bool	found;
 
-t_env	*env_into_list(char **envp)
-{
-	t_env	*env;
-	t_env	*tmp;
-	int		i;
+// 	found = false;
+// 	current = env;
+// 	while (current)
+// 	{
+// 		if (ft_strcmp(current->full_var, "_=/usr/bin/env") == 0)
+// 			found = true;
+// 		current = current->next;
+// 	}
+// 	if (found == false)
+// 	{
+// 		tmp = no_env("_=/usr/bin/env");
+// 		add_env_node(&env, tmp);
+// 		return ;
+// 	}
+// }
 
-	env = NULL;
-	i = 0;
-	if (!envp || !*envp)
-	{
-		tmp = no_env("_=/usr/bin/env");
-		add_env_node(&env, tmp);
-	}
-	else
-	{
-		while (envp[i])
-		{
-			tmp = create_env_node(envp, i);
-			add_env_node(&env, tmp);
-			i++;
-		}
-	}
-	add_underscore(env);
-	update_env(&env);
-	split_env(&env);
-	return (env);
-}
+// t_env	*env_into_list(char **envp)
+// {
+// 	t_env	*env;
+// 	t_env	*tmp;
+// 	int		i;
 
-char *join_name_var(t_env *env)
-{
-	char	*tmp;
-	char	*res;
+// 	env = NULL;
+// 	i = 0;
+// 	if (!envp || !*envp)
+// 	{
+// 		tmp = no_env("_=/usr/bin/env");
+// 		add_env_node(&env, tmp);
+// 	}
+// 	else
+// 	{
+// 		while (envp[i])
+// 		{
+// 			tmp = create_env_node(envp, i);
+// 			add_env_node(&env, tmp);
+// 			i++;
+// 		}
+// 	}
+// 	add_underscore(env);
+// 	update_env(&env);
+// 	split_env(&env);
+// 	return (env);
+// }
 
-	if (!env->value)
-		return (ft_strjoin(env->key, "="));
-	tmp = ft_strjoin(env->key, "=");
-	if (!tmp)
-		return (NULL);
-	res = ft_strjoin(tmp, env->value);
-	free(tmp);
-	if (!res)
-		return (NULL);
-	return (res);
-}
+// char *join_name_var(t_env *env)
+// {
+// 	char	*tmp;
+// 	char	*res;
 
-char **get_env(t_env *env)
-{
-	int			i;
-	const char	**cpy = ft_calloc(sizeof(char *), env_len(env) + 1);
-	t_env		*current;
+// 	if (!env->value)
+// 		return (ft_strjoin(env->key, "="));
+// 	tmp = ft_strjoin(env->key, "=");
+// 	if (!tmp)
+// 		return (NULL);
+// 	res = ft_strjoin(tmp, env->value);
+// 	free(tmp);
+// 	if (!res)
+// 		return (NULL);
+// 	return (res);
+// }
 
-	if (!cpy)
-		return (NULL);
-	current = env;
-	i = 0;
-	while (current)
-	{
-		if (is_equal(current->full_var))
-			cpy[i] = join_name_var(current);
-		else
-			cpy[i] = ft_strdup(current->key);
-		if (!cpy[i])
-			return (free_tab((char **)cpy), NULL);
-		current = current->next;
-		i++;
-	}
-	return ((char **)cpy);
-}
+// char **get_env(t_env *env)
+// {
+// 	int			i;
+// 	const char	**cpy = ft_calloc(sizeof(char *), env_len(env) + 1);
+// 	t_env		*current;
+
+// 	if (!cpy)
+// 		return (NULL);
+// 	current = env;
+// 	i = 0;
+// 	while (current)
+// 	{
+// 		if (is_equal(current->full_var))
+// 			cpy[i] = join_name_var(current);
+// 		else
+// 			cpy[i] = ft_strdup(current->key);
+// 		if (!cpy[i])
+// 			return (free_tab((char **)cpy), NULL);
+// 		current = current->next;
+// 		i++;
+// 	}
+// 	return ((char **)cpy);
+// }
 
 // -----------------------export_print.c---------------------------
 
-void	ft_swap_tab(char **a, char **b)
-{
-	char	*temp;
+// void	ft_swap_tab(char **a, char **b)
+// {
+// 	char	*temp;
 
-	temp = *a;
-	*a = *b;
-	*b = temp;
-}
+// 	temp = *a;
+// 	*a = *b;
+// 	*b = temp;
+// }
 
-char	**sort_env(char **tab, t_env *env)
-{
-	int			i;
-	const char	**sorted = ft_calloc(sizeof(char *), env_len(env) + 1);
+// char	**sort_env(char **tab, t_env *env)
+// {
+// 	int			i;
+// 	const char	**sorted = ft_calloc(sizeof(char *), env_len(env) + 1);
 
-	if (!sorted)
-		return (NULL);
-	i = 0;
-	while (i < env_len(env) - 1)
-	{
-		if (ft_strcmp(tab[i], tab[i + 1]) > 0)
-		{
-			ft_swap_tab(&tab[i], &tab[i + 1]);
-			i = 0;
-		}
-		else
-			i++;
-	}
-	i = 0;
-	while (tab[i] != NULL)
-	{
-		sorted[i] = ft_strdup(tab[i]);
-		if (!sorted[i])
-			return (free_tab((char **)sorted), NULL);
-		i++;
-	}
-	return ((char **)sorted);
-}
+// 	if (!sorted)
+// 		return (NULL);
+// 	i = 0;
+// 	while (i < env_len(env) - 1)
+// 	{
+// 		if (ft_strcmp(tab[i], tab[i + 1]) > 0)
+// 		{
+// 			ft_swap_tab(&tab[i], &tab[i + 1]);
+// 			i = 0;
+// 		}
+// 		else
+// 			i++;
+// 	}
+// 	i = 0;
+// 	while (tab[i] != NULL)
+// 	{
+// 		sorted[i] = ft_strdup(tab[i]);
+// 		if (!sorted[i])
+// 			return (free_tab((char **)sorted), NULL);
+// 		i++;
+// 	}
+// 	return ((char **)sorted);
+// }
 
-int	print_line(char *line, int fd)
-{
-	const char	*value = get_value_env(line);
-	const char	*name = get_key_env(line);
+// int	print_line(char *line, int fd)
+// {
+// 	const char	*value = get_value_env(line);
+// 	const char	*name = get_key_env(line);
 
-	if (fd == -1)
-		return (free((void *)name), free((void *)value), perror("msh"), 1);
-	if (!name)
-		return (free((void *)value), 1);
-	if (value)
-		ft_printf(fd, "declare -x %s=\"%s\"\n", name, value);
-	else
-		ft_printf(fd, "declare -x %s\n", name);
-	free((void *)name);
-	free((void *)value);
-	return (0);
-}
+// 	if (fd == -1)
+// 		return (free((void *)name), free((void *)value), perror("msh"), 1);
+// 	if (!name)
+// 		return (free((void *)value), 1);
+// 	if (value)
+// 		ft_printf(fd, "declare -x %s=\"%s\"\n", name, value);
+// 	else
+// 		ft_printf(fd, "declare -x %s\n", name);
+// 	free((void *)name);
+// 	free((void *)value);
+// 	return (0);
+// }
 
-int	print_export(char **sorted, t_msh *msh)
-{
-	int			i;
-	const int	fd = which_fd(msh);
+// int	print_export(char **sorted, t_msh *msh)
+// {
+// 	int			i;
+// 	const int	fd = which_fd(msh);
 
-	if (fd == -1)
-		return (perror("msh"), 1);
-	i = 0;
-	while (sorted[i])
-	{
-		if (print_line(sorted[i], fd))
-			return (1);
-		i++;
-	}
-	return (0);
-}
+// 	if (fd == -1)
+// 		return (perror("msh"), 1);
+// 	i = 0;
+// 	while (sorted[i])
+// 	{
+// 		if (print_line(sorted[i], fd))
+// 			return (1);
+// 		i++;
+// 	}
+// 	return (0);
+// }
 
 //---------------------export_utils.c-----------------------------------
